backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
psycopg2-binary==2.9.3
python-dotenv==0.19.0
boto3==1.26.137
opensearch-py==2.0.0
requests==2.28.1
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
nltk==3.6.2
spacy==3.1.3
transformers==4.11.3
torch==1.9.0
pinecone-client==2.2.1
fetchai-sdk==0.10.0
python-magic==0.4.24
gunicorn==20.1.0 